app/requirements.txt
Flask
psutil
pyyaml
requests